add non llm rel builder
